feat(app): add About settings tab, relocate version info and Updates
The version/identifier/build-date/ffmpeg row lived buried in Advanced,
and the Updates section lived in General separate from the version
display it relates to, with no link to the project repository anywhere.

Adds a dedicated About tab (last in the sidebar) that carries all of
it: the four identity rows plus a GitHub link, and the Updates section
moved wholesale from General. No behavior change to update checking —
this is a pure relocation. The new tab is added to the /ui/press
allowlist alongside the existing Speakers entry so it stays driveable
through the local automation API.
