+++ app/MeetingTranscriber/Tests/AboutSettingsViewTests.swift
